Prepare to publish on npm (code-dot-org only)
Simplifies the README to clarify that this is not the primary Piskel repository.

Overhauls the package.json file to clean up metadata and prepare to publish to npm as @code-dot-org/piskel, a package that is not a module, containing only the distribution build and metadata for this version of Piskel.

Adds an .npmignore file so that the .gitignore isn't used by default, which allows us to publish build products.