SmokeTest should always wait 10 seconds
Hard-codes the initial wait during smoke test to 10s.  Previously it was 10s on Travis but only 50ms on local.  I couldn't get it to pass on local with the shorter wait, but it passes on local if I increase the wait to 10s.  However, local runs a lot more tests and waiting 10s for every drawing test would be interminable.  Therefore, I've decided this is a special case, and we'll hard-code its wait time.

I wish I knew more about the root cause of needing to wait 10s for the frame to show up, but the inner workings of Casper+Phantom are still a mystery to me.