Replace cdo-message-api with CodeOrgMessageSerivce
Follows the structure of the Piskel application better by implementing the CDO message listener as a service (similar to the import service).  Removes the hack to add cdo-message-api.js to index.html, and loads the CodeOrgMessageService via the normal initialization process.  The new service has fewer global dependencies (window and the ImportService must be injected on initialization) and it uses the ImportService so it has less of its own import logic.