Export MessageType constant for use by external app.
Create an ExportedConstants.js file that gets included in the published package and is used as the "main" entry (so that a `require('piskel')` will load ExportedConstants.js).

Add a `MessageType` constant with our only message type, `LOAD_ANIMATION` defined.  Use it in CodeOrgMessageService.js and export it so that the outer application can use it when sending messages.