Export PiskelApi class instead of constants
Changes the way a parent application is supposed to interact with embedded Piskel.  Instead of expecting the parent application to implement all communication with Piskel, we provide a small wrapper object that knows how to send and receive messages to/from Piskel in an iframe, and provides methods for interacting with the embedded application.

Also starts hooking up reporting animation changes.